requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.9.0+cu111
-f https://download.pytorch.org/whl/torch_stable.html
torchvision==0.10.0+cu111
transformers==4.7.0
numpy==1.21.1
requests==2.26.0
matplotlib==3.4.2
Pillow==8.3.1
opencv-python==4.5.3.56
timm==0.4.12